pyproject.toml
[tool.poetry]
name = "gitlab-deployes"
version = "0.1.4"
description = "GitLab Deployer"
authors = ["Dmitry Vysochin <dmitry.vysochin@gmail.com>"]
license = "MIT"
homepage = "https://github.com/veryevilzed/gitlab-deployer"
repository = "https://github.com/veryevilzed/gitlab-deployer"
keywords = ["deploy", "tools"]
readme = "README.md"
classifiers = [
    "Topic :: Software Development :: Build Tools"
]

packages = [
    { include = "deployer" },
]
include = [
]


[tool.poetry.dependencies]
python = "^3.6"
Click = "^7.0"
requests = "^2.22"
python-gitlab = "^1.15"

[tool.poetry.scripts]
deployer = 'deployer.main:cli'

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


